Allow specifying resources in Op/Asset params list (#11322)
## Summary

Allows users to specify resource dependencies as parameters to an Asset
or Op function.

Resource dependencies are identified by type annotations which either
extend `ResourceDefinition` or are wrapped as `ResourceOutput[T]`.

The first pattern is useful for "new"-style resources, which are Python
classes that the user will interact with. In this case, the
`ResourceDefinition` and the output from the resource function are
identical. Here, a theoretical `S3Resource` implements
`ResourceDefinition` and provides various interactions:

```python
@asset
def my_asset(s3: S3Resource):
    with s3.open_bucket(...):
        .... 
```

The second pattern is useful for function-style resources, which may
output any arbitrary Python object. In this case, our `PostgresResource`
outputs a `DatabaseConnection` object:

```python
@op
def my_op(postgres: ResourceOutput[DatabaseConnection]):
    postgres.query(...)
```

## Implementation

For the prototype, this approach was chosen over an approach which would
use input names rather than annotations to identify which inputs are
resources and which are other op/asset dependencies for a few reasons:

- Clarity: By using annotations, it is immediately clear to the reader
which parameters are resources.
- Complexity: Asset dependencies are currently resolved immediately,
when the decorator is applied. It would require a significant rewrite to
delay this evaluation until the `Definitions` object containing the
asset is evaluated.


## Test Plan

Unit tests