change version placeholder 0+dev -> 1!0+dev (#11334)
### Summary & Motivation

Changes the placeholder version from `0+dev` to `1!0+dev`.

The motivation here is to guarantee that the placeholder version is always greater than any published version. This is accomplished by the `1!`, which is an "epoch" specifier (the implicit epoch on published versions is 0).

This means our local versions of packages can satisfy any lower-bounded dagster requirement. This recently became an issue with https://github.com/dagster-io/internal/pull/4420, which added the external package [dagster-hex (https://github.com/hex-inc/dagster-hex) as a dependency to some of our tests. `dagster-hex` depends on `dagster>1`, which was not being satisfied by our usual editable installs (which have version `0+dev`). Using `1!0+dev` fixes this problem.

### How I Tested These Changes

BK