RFC/WIP: Be more resilient to user code errors when dequeuing runs
Summary:
Right now if the user code server is unreachable while a run is being dequeued - e.g. you're using the default run launcher and the grpc server is down, or you're using cloud and the run is launched via the agent - the dequeue will immediately give up.

This is a work-in-progress solution that instead:
- Retries up to N times by re-enqueing the run
- Puts the last user code failure timestamp for that code location in a map to make it wait to try again for a while, so we don't race through the N retries right away

The other missing piece here would be to add a hook on each iteration to optionally check that all user code servers are unavailable - that would cover cases in cloud like the agent being in the middle of being upgraded for an extended period of time.