Add single dimension -> multidimension partition mapping (#10910)
Resolves https://github.com/dagster-io/dagster/issues/10722

Following https://github.com/dagster-io/dagster/pull/10909/files, this
PR adds a partition mapping from an upstream single-dimensionally
partitioned asset to a downstream multipartitioned asset. In order to
use this class, the upstream partitions definition must be a dimension
of the downstream asset.

The new `PartitionMapping` object also returns partition subsets instead
of ranges, enabling use cases where a downstream subset is not a valid
range e.g.:
```
MultiPartitionsDefinition(
    {"ab": StaticPartitionsDefinition(["a", "b"]), "12": StaticPartitionsDefinition(["1", "2"])}
)
```
has partition keys: ['1|a', '1|b', '2|a', '2|b']
If the upstream asset had partitions definition
StaticPartitionsDefinition(["a", "b"]), the downstream partition keys
subset should be {'1|a', '2|a'}, which can't be expressed as a range.