[dagit] Begin adding new graphql-codegen (#11411)
### Summary & Motivation

Apollo is deprecating its codegen CLI feature in favor of
https://the-guild.dev/graphql/codegen. Begin swapping out our old
codegen with the new.

- Use the recommended client preset to generate GraphQL types via `yarn
graphql-codegen`. The produced output can be seen in the `src/graphql`
directory. Namely, it produces:
- `graphql` functions that are bound to query documents, which can be
used at locations where queries are defined, instead of the `gql`
template literal. This allows us to make `useQuery` calls that do not
require annotation! Based on the supplied query, the variables and
output are all completely typechecked. (See `VersionNumber` for an
example.)
- Types for all schema fields and query outputs. These are similar to
our existing codegenned types, but they do *not* include the
intermediate per-field types for each query. This means we'll need to do
some refactoring in places where we reference nested types
(`MyQuery_someField_subFieldFoo`) to use fragments or similar.
- All types are located within `src/graphl`, so all GrahpQL type imports
will be from the same place (src/graphql/graphql).

In the codegen configuration, I am disabling fragment masking at this
point, since it adds more overhead to the refactor. We can add it later
if it seems like something we'll want.

I have updated `VersionNumber` to serve as an example. I'll make a
comment inline.

The recommendation (as noted in the codegen output file) is to use the
preset's provided Babel plugin to reduce bundle size, since all GraphQL
documents are output into the codegen file. We'll need to do this by
updating our CRA fork, since that's where our Babel config lives. For
now, our bundle size is a bit larger, like +20kb gzipped in the main
bundle.

The `yarn generate-graphql` command will now produce *both* old and new
types. When all of our GraphQL usage is moved over to the new API and
types, we can remove the old types entirely.

Next steps:

- Begin converting existing `gql` TTLs to `graphql()` calls and fix
types as we go.

### How I Tested These Changes

TS, lint, jest. Run same on dagit-cloud to verify that everything still
works.

Load Dagit, verify that the `VersionNumber` component behaves and
renders correctly.

Verify that lint does not run on `src/graphql` output, since that will
just be expensive and pointless. Verify that `__typename` values are
present on all types, and that enum values preserve `ALL_CAPS` keys.

Produce a prod build with stats, verify that the bundle hasn't exploded
in size due to the new codegen output.