Branching I/O Manager (#11315)
### Summary & Motivation

Stemming from a conversation with Daniel Gafni.

A common problem in ML pipelines (and data pipelines in general) is that you want to develop on upstream production assets but write to a dev or branch environment during iterative development. Sometimes the source data for an ML model is the result of expensive and long-running computations and rerunning those is very expensive. You want to be able to use those production assets and then fork storage into a staging environment for the assets one wants to iterate on.

Data warehouses like Snowflake support this natively with copy-on-write CLONE features. However in the generalized case one needs to control this with software. This is what this PR seeks to enable.

It works by leveraging Dagster's metadata to "mark" certain materializations as created in dev and then creating a BranchingIOManager which composes two i/o managers, one for the branch environment and one for the parent environment. The handle_output method of BranchingIOManager writes to the branch environment unconditionally, and creates an AssetMaterialization with an "io_manager_branch" entry (is configurable) that marks it as "dev" (again, configurable). The `load_input` checks for a previous AssetMaterialization marked in that branch. If that exists, it reads from the branch. Else it reads from parent.

This also leverages the new Definitions API to make this more tractable, as it relies on object references to created IOManagers, rather than @io_manager-decorated functions. We can simply pass in created resources rather than relying on the `required_resource_key` system.

### How I Tested These Changes

Included tests

Also developed https://github.com/schrockn/dagster-script-to-assets/blob/working-branch/hn_dagster.py against this