Fix tick graphql resolver creating an error of the wrong type
Summary:
I stumbled on a bit of a landmine on a separate PR by adding logic to GraphenePythonError that assumes that it's a GraphenePythonError - normally a valid assumption, but if you pass a namedtuple where Graphene is expecting a GrapheneObjectType, it appears that it coerces it into the object without ever calling the constructor. So we were calling GraphenePythonError methods on a SerializableErrorInfo.

I'll audit other callsites of GraphenePythonError for places we might have been doing this, but this is the only one i've found so far.

Test Plan: New test that would fail in master