Add StructuredIOManagerAdapter (#11383)
### Summary & Motivation

Similar to StructuredResourceAdapter, add StructuredIOManagerAdapter. This allows a user to take an existing @io_manager or IOManagerDefinition and use it off the shelf using "new-style" resources.

Imagine you have an existing IOManager class:

```python
    @io_manager(config_schema={"a_config_value": str})
    def an_io_manager(context: InitResourceContext) -> AnIOManagerImplementation:
        return AnIOManagerImplementation(context.resource_config["a_config_value"])
```

You can adapter it to the new style as so:

```python
    class AdapterForIOManager(StructuredIOManagerAdapter):
        a_config_value: str
        @property
        def wrapped_io_manager(self) -> IOManagerDefinition:
            return an_io_manager
```

And then use it instead of a `.configured` call.

```python
    defs = Definitions(
        assets=[an_asset],
        resources={"io_manager": AdapterForIOManager(a_config_value="passed-in-configured")},
    )
```

### How I Tested These Changes

BK