Support pydantic aliases in config schema mapping (#11389)
### Summary & Motivation

Discovered this issue while trying to put a config field `schema` into a new-style config class. Pydantic requires that you alias this, so updated the mapping to respect that alias field. This makes it so that you refer to the alias in config space, but the raw name at runtime.

For example.

```
    class ConfigWithSchema(Config):
        schema_: str = pydantic.Field(alias="schema")
```

is equivalent to the config schema: `{"schema": dagster.Field(StringSource)}`

In an op it looks like:

```
    @op
    def an_op(context, config: ConfigWithSchema):
        # use the raw property in python space
        assert config.schema_ == "bar"
        # use the alias in config space
        assert context.op_config == {"schema": "bar"}
  ```

### How I Tested These Changes

BK