[dagit] Remove old GraphQL Codegen (#11474)
### Summary & Motivation

Entirely remove Apollo codegen from dagit OSS.

The bulk of the changes here (apart from all the codegen stuff being
deleted) is the replacement of all fragment `gql` documents with
`graphql()` calls, and the update of downstream type imports and
annotations.

### How I Tested These Changes

yarn generate-graphql, verify that all old codegen is gone and not
recreated.

TS, lint, Jest. Do the same in dagit-cloud to ensure that dagit-cloud
master is unaffected.