Typehint and validate tags and status params to create_run (#11456)
### Summary & Motivation

Typehint and check the tags and status parameters to create_run.

Tags in callsites comes from sources that are both Mapping[str, str] and
Mapping[str, object]. Doing a validation pass to ensure that they are in
fact str to str.

Status was relying on the fact that all callsites to
create_run_for_pipeline were not respecting the typehint on its status
parameter and were passing the enum instead.

### How I Tested These Changes

BK