Possible fix for run monitoring issues: consider the run worker unhealthy is the job has no active pods but the run has started
Summary:
Not sure if this is too much of a bandaid, but a user reported an issue where run worker crash recovery didn't kick in for a STARTED run - the speciif

Need to figure out a good way to test this.. it probably involves unit tests vs. an e2e test.