only count materializations within backfill (#11506)
### Summary & Motivation

Asset backfills look at what asset partitions have already been
materialized to decide what partitions need to be submitted.

Prior to this change, the backfill code would look at all
materializations of the asset, independent of whether they were part of
the backfill. This meant that any materialization prior to the backfill
would stop the backfill from submitting runs.

This change filters to only materializations that were part of the
backfill. It functions by adding a backfill tag to the relevant
materializations

An alternative way to implement this would have been to look for the
backfill tag on the runs, instead of on the asset materializations. This
would have required joining the event log with the run tags table.

### How I Tested These Changes

Added a test that fails before the change