Add experimental examples directory (#11505)
### Summary & Motivation

We'd like to port existing examples to experimental APIs as well as create new examples that use experimental APIs. We don't want users copying and pasting these examples or to fully support them. So this proposes to create a new experimental folder in examples for this purpose.

My immediate goal is to create a copy of `project_fully_featured` and then port it to structured resources and config.

### How I Tested These Changes

BK

Ran commit that touched core and ran it to confirm that EXAMPLES_TO_IGNORE worked as intended

See https://buildkite.com/dagster/dagster/builds/44511 (canceled). Reran @ https://buildkite.com/dagster/dagster/builds/44514