update black 22.3.0 -> 22.12.0 and enable --preview (#10900)
### Summary & Motivation

Updates `black` from 22.3.0 to 22.10.0 and enables `--preview` mode.

The version bump gives a small performance boost and bugfixes. Enabling
`--preview` mode uses tweaks to the black style that are highly likely
to be included in next major release, it's kind of like a conservative
"enable-experimental-features". The rule changes that make up 99% of
changes in this changeset include:

- Formatting of long string literals. There are a lot of these in the
repo that had to be manually adjusted before, so this is a win.
- Parens around return annotations now managed
- Deletion of blank lines at the start of a code block
- Several cases where extraneous parentheses are removed (awaited
objects, with statements)

Comprehensive list is in [Black's
changelog](https://black.readthedocs.io/en/stable/change_log.html).

### How I Tested These Changes

Existing BK tests.

Internal companion PR: https://github.com/dagster-io/internal/pull/4323