[dagit] Use cache-and-network as default watchQuery fetch policy (#11523)
### Summary & Motivation

Modify our ApolloClient defaults to use `cache-and-network` as the
default fetch policy on watched queries, instead of `cache-first`.

This should ideally resolve issues where the backend has changed (e.g. a
code change for a job) and needs to be re-queried. We use
`cache-and-network` pretty aggressively already to get this behavior,
and my hunch is that we don't actually want `cache-first` very often
since the backend is always going to be a more appropriate source of
truth for anything that needs requerying.

I have gone through our existing `useQuery` calls to add `cache-first`
as the fetch policy on any that haven't specified it, though we should
audit these to see which of them should actually be `cache-and-network`.
I've added comments to each.

I have also removed any `fetchPolicy: 'cache-and-network'` settings from
`useQuery` since this will now be the default.

The end result is that all queries should keep the exact same behavior
they currently have (to avoid unexpected breakages) and we can go
through and fix any queries that are incorrectly configured.

### How I Tested These Changes

Load Dagit and sanity check that all queries are still performed as
expected.

Verify fix for the issue @gibsondan ran into, now that the query in
`PipelineExplorerRoot` now uses the `cache-and-network` default:

- Load Dagit. Navigate to Deployments page.
- Change `fails_job` to have another op in Python.
- Kill the dagit backend, then restart it.
- Wait until Dagit says that definitions are reloaded.
- Navigate to `fails_job` in Dagit, verify that it now has the op that I
added. This is because navigating to the job page led to a fresh query
of the backend, as expected with `cache-and-network`.