[dagit] Track most recent preference for computed log type view (#11527)
### Summary & Motivation

Use localStorage to track the preference between `stdout` and `stderr`
for which tab the user will see first when switching to compute logs on
a Run page.

- With no localStorage value set, `stdout` will be the default.
- If the user clicks on either `stdout` or `stderr` tab, track that
value.
- If the user then navigates to a run and clicks the button to view
compute logs, that tab will be the one that shows first.
- If the user changes tabs, update the tracked value.
- If the user navigates to a run with a `logType` query param (e.g.
`logType=stderr`), show that compute log tab but do not set the user's
preference locally.

### How I Tested These Changes

Verify the bullet points above when viewing a run. Load a run, switch to
compute logs. Verify that `stdout` is the default. Click `stderr`,
reload the page with no `logType` query param, switch to compute logs.
Verify that the `stderr` tab shows first.