refactor(dbt-cloud): remove v2 suffix (#11586)
### Summary & Motivation
The version of the dbt Cloud API should not be leaked in the class name,
especially since we don't expose an actual client object over all the v2
endpoints. In the future, this will be migrated to the v3 api, or even
live in a state between v2 and v3, so remove the version from the name
to avoid confusion.

### How I Tested These Changes
bk