[eslint-config] Enable ignoreExternal on no-cycle rule (#11632)
### Summary & Motivation

Repair eslint speed by enabling `ignoreExternal` on `import/no-cycle`.
Something is taking a dramatically long time in lint on this rule
specifically. I'm not sure what it is exactly, but it seems likely that
it's related to the recent GraphQL codegen change. This rule change
repairs the problem and gets lint back to a sane run time.

### How I Tested These Changes

`TIMING=1 yarn lint`, verify that lint speed is back to normal.