use run tags instead of asset tags to filter materializations within backfill (#11555)
### Summary & Motivation

This is a reversal of the approach taken in
https://github.com/dagster-io/dagster/pull/11506.

I realized a problem with that earlier approach: it depends on user code
being upgraded. I.e. if someone running a Dagster 1.1.8 daemon runs a
backfill against user code that depends on Dagster 1.1.7 or earlier,
then the asset materialization events that are created as part of that
backfill won't have the expected backill tag on them.

As a consequence, if that backfill has some asset partitions that can’t
be run until other materializations complete, then the backfill will
fail to terminate. This should only be a problem right now in very rare
situations, but will be more of a problem when we remove restrictions
that prevent users from launching backfills across different code
locations and across different partitions definitions.

### How I Tested These Changes

The expected behavior is covered by existing test coverage. In an ideal
world, we would write a test that makes sure backfills kicked off on
later versions of host code complete successfully on earlier versions of
user code, but I can't think of a way to do that that avoids a bunch of
very heavy lifting.