[dagit] Stop foreign node links from creating sprawling asset graphs (#11588)
### Summary & Motivation

This is a fix for one of the most pressing issues @braunjj identified in
his asset graph explorations -- we weren't capping the length of
external graph links on the Asset Graph, so they could easily result in
horizontal expansion of the graph.

This PR limits them to 10 characters, which is enough that two can be
shown side-by-side above an asset node without increasing the graph
width. We could potentially just show the icon or redesign these futher,
but this should be a good incremental improvement.

In this PR I also simplified some of the asset node sizing. Rather than
trying to vary the size of the boxes a bit based on their names, we make
them the same width all the time. This results in better renderings
because it's more likely that the assets fit into a nice grid. It also
allows us to fix a bug in the "minimal" rendering where occasionally the
labels would overflow their containers.

### How I Tested These Changes

Before:
![Screen Shot 2023-01-09 at 2 34 33
PM](https://user-images.githubusercontent.com/1037212/211403295-12b798f1-d9a6-4c78-9676-f1de0c179ff9.png)

After:
![Screen Shot 2023-01-09 at 2 34 20
PM](https://user-images.githubusercontent.com/1037212/211403292-9844b1df-1b54-4a5b-8b63-431705a9d691.png)

Co-authored-by: bengotow <bgotow@elementl.com>