Shut down the daemon faster when it is interrupted (#11657)
Summary:
we shut down the daemon by setting an event that each daemon thread
checks to terminate cleanly. Sometimes a particular daemon is in the
middle of a long sleep when this happens, making the daemon take forever
to spin down. Pass the shut down event through to the daemons so that
instead of calling time.sleep, they can call shutdown_event.wait

### Summary & Motivation

### How I Tested These Changes