dagit => dagster dev in docs
Summary:
Point people to `dagster dev` instead of `dagit` in the docs, except when we're specifically talking about deploying to production.

A separate PR will add a new section that explains more about `dagster dev` in a "Running Dagster Locally" guide - this PR is more rote and is pretty much just find-replacing things.