feat(dbt-cloud): cache compile run id as env var for job (#11691)
### Summary & Motivation
When generating software-defined assets from a dbt Cloud project, most
of the latency comes from compiling the project. Rather than compiling
the dbt project in an ad-hoc manner, only compile it when either:

- the project changes (via git commit), or when
- the dbt Cloud job commands change

If one of those conditions are met, then compile the project. And save
the reference/overwrite the reference to the compilation.

### How I Tested These Changes
pytest, local