Heartbeat once at the beginning of every interval daemon
Summary:
We're seeing a report from a user about the backfill daemon giving up due to a lack of heartbeats - but curiously it is happening well before our default tolerance limit of 5 minutes. I suspect that what is happening is that the backfill daemon does something expensive right at the beginning before it yields anything - so the only heartbeat we have is super old and the tolerance doesn't really matter.

To resolve this, do an initial yield / potential heartbeat at the beginning of every daemon. So now we will have a baseline and will still catch it if the daemon hangs.