Log to daemon output when running the schedule code fails or times out
Summary:
We write the tick as a failure but we don't log to the daemon output - making it hard to understand what's going on when schedules are timing out or otherwise failing. add a lot in the same place as the sensor daemon.

While we're in here, change some .errors to .exceptions since that's the more appropriate level here