Move pyproject.toml => origin generation to a WorkspaceLoadTarget class so that it can be reloaded in dagit
Summary:
This gives us the nice property that you can change the contents of the pyproject.toml file and hit "Reload All" in dagit without needing to restart it to pick up the changes. Also clears the deck for returning multiple modules which I remember was requestsed.

Test Plan: BK, sanity check with a local dagit load