Fix multidimensional partition backfills (#11788)
This PR rolls up a couple of fixes into one.

- Currently, multidimensional asset backfills do not update partition
status in Dagit. This is because recent changes to backfill logic did
not add the corresponding per-dimension tags to the requested runs. This
PR fixes this.
- The run status column of the backfill page displays as empty. This is
because:
- Multidimensional asset runs do not write a partition to the runs table
  - Minor JS bug (fix enclosed in PR)

This PR amends multidimensionally partitioned runs to also have the
`partition` tag, which is the multidimensional partition key e.g.
`a|2022-01-01`, subsequently writing this tag to the partition column of
the runs table. Also updates the asset backfill logic to also add the
multipartitions tags to asset backfill runs so multidimensional runs
also contain per-dimension tags.