Fix for backoff logic incorrectly storing state (#11848)
Summary:
User reported an error with our backoff logic hitting a maximum python
int limit. This is likely because we were initializing the default
backoff generator once and keeping it running forever. This fixes that.

One question is whether we need to do anything to clean up the old
generator as well.

### Summary & Motivation

### How I Tested These Changes